seriell_program/requirements.txt
pyserial
neovim
